requirements.txt
flask==3.1.0
requests==2.32.3
Pillow==11.1.0
python-dotenv==1.0.1
werkzeug==3.1.3
gunicorn==21.2.0
